requirements.txt
flask
python-dotenv
mysql-connector-python
gunicorn
mecab-python3
